Testing: Rename some files and fix tests
