+++ spyder/widgets/tests/test_findinfiles.py
