+++ spyder/widgets/variableexplorer/tests/test_utils.py
